% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_assignment.R
\name{initialize_assignment}
\alias{initialize_assignment}
\title{Initialize assignment}
\usage{
initialize_assignment(assignment)
}
\arguments{
\item{assignment}{}
}
\description{
Initialize assignment
}
