% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assignment_for_courseworks.R
\name{process_assignment_for_courseworks}
\alias{process_assignment_for_courseworks}
\title{Process assignment for courseworks}
\usage{
process_assignment_for_courseworks(assignment = NULL)
}
\arguments{
\item{assignment}{character; homework to process (e.g. "HW1")}
}
\description{
For a single assignment, process the google spreadsheet to submit assignment grades to courseworks
}
