% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assignment_for_email.R
\name{process_assignment_for_email}
\alias{process_assignment_for_email}
\title{Process assignment for email}
\usage{
process_assignment_for_email(sheet)
}
\arguments{
\item{assignment}{character; homework to process (e.g. "HW1")}
}
\description{
For a single assignment, process the google spreadsheet to distribute grades to students via email
}
