% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email_to_students.R
\name{send_emails_to_students}
\alias{send_emails_to_students}
\title{Send emails to students}
\usage{
send_emails_to_students(sheet)
}
\arguments{
\item{sheet}{}
}
\value{

}
\description{
Send emails to students
}
